Split nparse QApplication class from the nparse application script
